fixed up mis-merges.
Ensure that David's use of rows/cols is honored (I/Tony will put
back in the statistical approach of cases/vars later)

removed unittests-dataframes as there is a bug to fix.

move functions requiring full availability of dataframe stuff to a last-loaded file.

make import.lisp compile, need to fix this up.

added dfcolumn / dfrow to exported symbols for dataframes.  Maybe we need to use "DF" as a package nickname, so that DF:name is the call?)

packages.lisp rectified for the above changes.

things now compile again (but might not work).

Signed-off-by: aj rossini <blindglobe@gmail.com>